Restructured the repository
Extracted the `infra/config.nix` file into modules for each component
of our infrastructure. We minimize as much as possible using the "\$"
syntax to delay calls to Terraform in favor of importing Nix outputs
to pass variables between modules. We will add some of the specific
services referenced in `infra/default.nix` in the upcoming commits.