Fix /var/secrets/google/key.json file not found error
In order for this file to exist in the volume, we needed a key
in the kubernetes secret data object to be "key.json", but our
keys were all of the JSON keys in the generated file, so each
was previously split into a separate volume file.

We also removed the unnecessary call of the Terraform function
base64decode by using the `binary_data` attribute on the
`kubernetes_secret` resource.