deleted a bunch of crap and almost set up local kubernetes authentication in terraform
we don't need any of the google stuff anymore because we are pivoting
away from large tech conglomerates with their crappy proprietary
software that steals your data.

i've also included a refactor with better design that uses nix modules
properly.

this mostly seems to work but there's currently a problem with the
kubernetes server only being available as root on our server, so we need
to fix that first before confirming that the terraform deployments will
work correctly with the kubernetes provider