finalize traefik module
finished converting the yaml files into a nix config.
i like the design of extracting out common variables instead of
accessing values from `config` unless they are from another module or an
option. also this is the first time we have created anything with this
rewrite so the tfstate files are excluded. and no more vim :) but still
using macos unfortunately