finally deployed forgejo correctly
we had to manage a PV and PVC for the Forgejo module (and Postgres
database) by ourselves because we are using a Local Storage class (i.e.
a path mounted on the server node), which does not support Dynamic
Provisioning. to achieve this, we created resources for each, allowing
the volume to attach to the correct node with the
`kubernetes.io/hostname` label.

while this solved the persistence issue with the application itself, the
Postgres dependency for some reason, did not respond to the PVC
configuration and we filed a [bug report](https://codeberg.org/forgejo-contrib/forgejo-helm/issues/162) about this in the Forgejo Helm
chart repository. Ultimately, we decided to manage the database
separately using configuration options that did ultimately work.