add k3d terranix module to create cluster
this terraform configuration is the same as running:

```sh
k3d cluster create bunkbed \
--k3s-arg "--disable=traefik@server:0" \
--port "4443:443@loadbalancer" \
--agents 3
```

you have to explicitly update the kube/config file with this.
we make the traefik helm release only dependent on the cluster resource
because it is the target of the first stage of the entire setup. because
the other services are dependent on traefik, we just need traefik to
require the cluster