fix host map and ssh connections when working remote
we kill the ssh process when running all because each separate call to
the script will handle the forwarding separately.

only the sirver-k8s context actually needs a host map set up because we
use the api.kube host on the server for this cluster configuration. the
host defaults to localhost for sirver-k3s

shellcheck recommended the parameter substitution instead of sed also