deploy immich with https support
the immich chart forced us to create a PVC, which was mostly
straightforward except that we needed to set `wait_until_bound = false`
because apparently the PVC only finishes deploying in terraform once it
can be bound, which, based on the storage class WaitForFirstConsumer
setting, expects a pod to be requesting it. the pod only requests it
though once the helm release is started, but because of the terraform
dependency on the PVC, it couldn't start.