[wip] deploy jellyfin
Jellyfin unfortunately is not designed to horizontally scale, in part
because of a single database lock, but it is helpful to have it deployed
through Kubernetes alongside everything else so that we only configure
within one ecosystem. This doesn't work currently because I think not
all of the necessary ports are exposed, but I'm honestly not sure. Due
to the lack of horizontal scalabilty of the application, there was no
official Helm chart, and all of the other ones I could find had
custom-built Docker images that obscured what they were adding to the
baseline functionality of the original Jellyfin.